% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vin.R
\name{vin_decode}
\alias{vin_decode}
\title{Decode a VIN (Vehicle Identification Number)}
\usage{
vin_decode(vin, year = NULL)
}
\arguments{
\item{vin}{17-character Vehicle Identification Number}

\item{year}{Optional model year to improve accuracy}
}
\value{
tibble: variable, value, variable_id, value_id
}
\description{
Decode a VIN (Vehicle Identification Number)
}
