% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vin.R
\name{vin_models}
\alias{vin_models}
\title{Get models for a make and year}
\usage{
vin_models(make, year)
}
\arguments{
\item{make}{Make name (e.g. "toyota", "ford", "honda")}

\item{year}{Model year (e.g. 2024)}
}
\value{
tibble: make_id, make_name, model_id, model_name
}
\description{
Get models for a make and year
}
